requirements.txt
PyPDF2>=3.0.0
python-docx>=0.8.11
sentence-transformers>=2.2.0
numpy>=1.24.0
